Restyled index page, expanded nav functionality
Needed a way to toggle a 'sticky' class on the top navigation to keep
it anchord to the top of the viewport when the user scrolls above or
below the nav's initial starting point. Went with an event listener in
Javascript.

A "scroll-to-top" button was added to the top navigation. This button
should return the user to the top of the page when clicked. Briefly
considered using an anchor tag, but added functionality via Javascript.